requirements.txt
google-genai>=1.0.0
openai>=1.40.0
python-dotenv>=1.0.0
Pillow>=10.0.0
